+++ tests/test_ZigzagDCTcoefficientsOrder.cpp
@@ -3,26 +3,26 @@
 
 TEST(ZigzagDCTcoefficientsOrder, test_encoding)
 {
-    std::vector<std::vector<std::vector<int>>> imageBlock = {{{1, 2, 6, 7},
-                                                              {3, 5, 8, 13},
-                                                              {4, 9, 12, 14},
-                                                              {10, 11, 15, 16}}};
+    std::vector<std::vector<std::vector<float>>> imageBlock = {{{1.0f, 2.0f, 6.0f, 7.0f},
+                                                                {3.0f, 5.0f, 8.0f, 13.0f},
+                                                                {4.0f, 9.0f, 12.0f, 14.0f},
+                                                                {10.0f, 11.0f, 15.0f, 16.0f}}};
     ZigzagDCTcoefficientsOrder encoder;
     auto zigzagOrder = encoder.ZigZagOrderImageBlocks(imageBlock);
-    std::vector<std::vector<int>> expected = {{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16}};
+    std::vector<std::vector<float>> expected = {{1.0f, 2.0f, 3.0f, 4.0f, 5.0f, 6.0f, 7.0f, 8.0f, 9.0f, 10.0f, 11.0f, 12.0f, 13.0f, 14.0f, 15.0f, 16.0f}};
     ASSERT_EQ(zigzagOrder, expected);
 }
 
 TEST(ZigzagDCTcoefficientsOrder, test_decoding)
 {
     ZigzagDCTcoefficientsOrder encoder;
-    std::vector<std::vector<int>> diagonalised = {{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16}};
+    std::vector<std::vector<float>> diagonalised = {{1.0f, 2.0f, 3.0f, 4.0f, 5.0f, 6.0f, 7.0f, 8.0f, 9.0f, 10.0f, 11.0f, 12.0f, 13.0f, 14.0f, 15.0f, 16.0f}};
     auto Imageblocks = encoder.deZigZagOrderImageBlocks(diagonalised);
 
-    std::vector<std::vector<std::vector<int>>> expectedImageBlock = {{{1, 2, 6, 7},
-                                                                      {3, 5, 8, 13},
-                                                                      {4, 9, 12, 14},
-                                                                      {10, 11, 15, 16}}};
+    std::vector<std::vector<std::vector<float>>> expectedImageBlock = {{{1.0f, 2.0f, 6.0f, 7.0f},
+                                                                        {3.0f, 5.0f, 8.0f, 13.0f},
+                                                                        {4.0f, 9.0f, 12.0f, 14.0f},
+                                                                        {10.0f, 11.0f, 15.0f, 16.0f}}};
 
     ASSERT_EQ(Imageblocks, expectedImageBlock);
 }
